# Improve performance of drawing bugpath