I improved the performance of drawing bugpath by using CodeMirror `operation` function: http://codemirror.net/doc/manual.html#operation. 

Bubbles (num) | Lines (num) | Current time (ms) | Improved time (ms)
-- | -- | -- | --
25 | 22 | 149.45 | 77.65
31 | 28 | 4102.05 | 2242.63
35 | 31 | 4954.74 | 2797.91
43 | 38 | 6108.59 | 3304.29
47 | 42 | 6716.03 | 3703.16
52 | 46 | 7571.04 | 4196.20
56 | 49 | 7375.35 | 4190.46
54 | 48 | 8973.17 | 4487.09

![cc_bugpath_drawing](https://user-images.githubusercontent.com/6695818/29413771-db80b8a8-835d-11e7-9c0d-4925fc037dfa.png)